+++ plugins/node/template/addons/oauth/src/modules/oauth/index.ts
@@ -1,0 +1,3 @@
+export * from "./oauth.schema";
+export * from "./oauth.service";
+export { default as oauthRoutes } from "./oauth.routes";